+++ src/include/cvc5_private.h
@@ -17,9 +17,6 @@
 #ifndef CVC5_PRIVATE_H
 #define CVC5_PRIVATE_H
 
-#if !(defined(__BUILDING_CVC5LIB) || defined(__BUILDING_CVC5LIB_UNIT_TEST))
-#  error A private cvc5 header was included when not building the library or private unit test code.
-#endif
 
 #include "base/cvc5config.h"
 #include "cvc5_public.h"
