+++ src/parser/CMakeLists.txt
@@ -33,6 +33,8 @@ set(libcvc5parserapi_src_files
 set(libcvc5parser_src_files
   commands.cpp
   commands.h
+  command_status.cpp
+  command_status.h
   flex_input.cpp
   flex_input.h
   flex_lexer.cpp
