+++ src/parser/command_status.cpp
@@ -1,0 +1,94 @@
+/******************************************************************************
+ * Top contributors (to current version):
+ *   Andrew Reynolds, Tim King, Morgan Deters
+ *
+ * This file is part of the cvc5 project.
+ *
+ * Copyright (c) 2009-2023 by the authors listed in the file AUTHORS
+ * in the top-level source directory and their institutional affiliations.
+ * All rights reserved.  See the file COPYING in the top-level source
+ * directory for licensing information.
+ * ****************************************************************************
+ *
+ * Command status utility
+ */
+
+#include "parser/command_status.h"
+
+#include <exception>
+#include <iostream>
+#include <iterator>
+#include <sstream>
+#include <utility>
+#include <vector>
+
+#include "base/check.h"
+#include "base/modal_exception.h"
+#include "base/output.h"
+#include "expr/node_manager.h"
+#include "options/io_utils.h"
+#include "options/main_options.h"
+#include "options/options.h"
+#include "options/printer_options.h"
+#include "options/smt_options.h"
+#include "parser/api/cpp/symbol_manager.h"
+#include "printer/printer.h"
+#include "proof/unsat_core.h"
+#include "util/smt2_quote_string.h"
+#include "util/utility.h"
+
+using namespace std;
+using namespace cvc5::parser;
+
+namespace cvc5::parser {
+
+const CommandSuccess* CommandSuccess::s_instance = new CommandSuccess();
+const CommandInterrupted* CommandInterrupted::s_instance =
+    new CommandInterrupted();
+
+std::ostream& operator<<(std::ostream& out, const CommandStatus& s)
+{
+  s.toStream(out);
+  return out;
+}
+
+ostream& operator<<(ostream& out, const CommandStatus* s)
+{
+  if (s == NULL)
+  {
+    out << "null";
+  }
+  else
+  {
+    out << *s;
+  }
+  return out;
+}
+
+void CommandSuccess::toStream(std::ostream& out) const
+{
+  internal::Printer::getPrinter(out)->toStreamCmdSuccess(out);
+}
+
+void CommandInterrupted::toStream(std::ostream& out) const
+{
+  internal::Printer::getPrinter(out)->toStreamCmdInterrupted(out);
+}
+
+void CommandUnsupported::toStream(std::ostream& out) const
+{
+  internal::Printer::getPrinter(out)->toStreamCmdUnsupported(out);
+}
+
+void CommandFailure::toStream(std::ostream& out) const
+{
+  internal::Printer::getPrinter(out)->toStreamCmdFailure(out, d_message);
+}
+
+void CommandRecoverableFailure::toStream(std::ostream& out) const
+{
+  internal::Printer::getPrinter(out)->toStreamCmdRecoverableFailure(out,
+                                                                    d_message);
+}
+
+}  // namespace cvc5::parser
